+++ bin/index.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node --experimental-json-modules
 
-import { cli } from '../build/index.mjs';
+import { cli } from '../build/index.js';
 
 cli(process.argv);