+++ src/enums.ts
@@ -1,0 +1,4 @@
+export enum Templates {
+  JavaScript = 'javascript',
+  TypeScript = 'typescript',
+}