+++ templates/javascript/src/pages/home.js
@@ -2,14 +2,12 @@ import './home.css';
 import edelweiss from './edelweiss.svg';
 import { html } from '@prostory/edelweiss';
 
-export default function homePage() {
+export function home() {
   return html`
     <main class="main">
       <img class="logo" src="${edelweiss}" alt="Logo" />
       <nav class="links">
-        <a class="link" href="https://github.com/YevhenKap/edelweiss#edelweiss"
-          >Documentations</a
-        >
+        <a class="link" href="https://yevhenkap.github.io/">Documentation</a>
         <a class="link" href="https://github.com/YevhenKap/edelweiss">GitHub</a>
       </nav>
     </main>