+++ templates/javascript/src/routes.js
@@ -1,10 +1,10 @@
-import homePage from './pages/home';
+import { home } from './pages/home';
 import { Router } from '@prostory/edelweiss';
 
 Router.add({
   path: '/',
   container: '#app',
   view() {
-    return homePage();
+    return home();
   }
 });