+++ templates/typescript/src/service_worker/service_worker_controller.d.ts
@@ -1,2 +1,4 @@
 /** Register service worker, if it is supported by browser. */
 export function registerServiceWorker(): void;
+/** Unregister service worker, if it is supported by browser. */
+export function unregisterServiceWorker(): void;